refactor: Todo型をインターフェースに移行し、不要なコードを削除
