+++ front/src/interfaces/index.ts
@@ -1,4 +1,4 @@
-export type Todo = {
+export interface Todo {
   id: number;
   title: string;
   content: string;
